Merged PR 102096: Improved logging for imodel-schema-validator
Previously, the tool was dependent on the log files coming from schema-validator and schema-comparer tools where logs for comparison and bis rules validation were separate.

Depending upon the need of improved logging, now we log bis rules and comparison validation results for each schema in a file. We also log the summary and results of all 4 validations into a separate file.